pyproject.toml
[project]
name = "try-fasthtml"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Michael Booth", email = "michael@databooth.com.au"},
]
dependencies = [
    "python-fasthtml>=0.2.1",
    "duckdb>=1.0.0",
    "pandas>=2.2.2",
    "toml>=0.10.2",
    "markdown>=3.6",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "Apache 2.0"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
lint = [
    "black[jupyter]>=24.4.2",
]
